changing type of stacking zone data to int. Fixes #290. multipoint registration still has some sign issues in shifts, the usual.

+++ src/gui/zones.h
@@ -9,7 +9,7 @@ int get_side(const stacking_zone *zone);
 int point_is_inside_zone(int px, int py, const stacking_zone *zone);
 gboolean zone_is_too_close(int x, int y, int size);
 
-void add_stacking_zone(double x, double y, double half_side);
+void add_stacking_zone(int x, int y, int half_side);
 void remove_stacking_zones();
 
 #endif
